Extract inbox and terminal viewport hooks
